chore: implement flight controller

+++ ticket/src/main/java/rsoi/lab2/ticket/repoistory/TicketRepoistory.java
@@ -2,6 +2,6 @@ package rsoi.lab2.ticket.repoistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import rsoi.lab2.ticket.model.Ticket;
 
-public interface TicketRepoistory extends JpaRepository<Ticket, Integer> {
+public interface TicketRepoistory extends JpaRepository<Ticket, Long> {
 
 }
